AddBook and AddComment controllers were removed
