+++ dz14/src/main/java/ru/otus/dz14/controller/BookController.java
@@ -39,10 +39,6 @@ public class BookController {
         return "redirect:/books";
     }
 
-//    @RequestMapping(
-//            value = {"/books/add"},
-//            method = RequestMethod.POST
-//    )
     @PostMapping("/books/add")
     public String saveBook(
             @ModelAttribute("bookDto") BookDto bookDto
@@ -53,10 +49,6 @@ public class BookController {
         return "redirect:/books";
     }
 
-//    @RequestMapping(
-//            value = {"/books/add/{id}"},
-//            method = RequestMethod.POST
-//    )
     @PostMapping("/books/add/{id}")
     public String updateBook(
             Model model,
@@ -67,5 +59,25 @@ public class BookController {
                 bookDto.getAuthorName(), bookDto.getAuthorLastName(),
                 bookDto.getGenre());
         return "redirect:/books";
+    }
+
+    @GetMapping("/addbook")
+    public String addBookPage(Model model) {
+        BookDto bookDto = new BookDto();
+        model.addAttribute("bookDto", bookDto);
+        return "addbook";
+    }
+
+    @GetMapping("/addbook/edit")
+    public String editBookPage(
+            @RequestParam("id") Integer id,
+            Model model
+    ) {
+        Optional<Book> bookOptional = libraryService.findBookById(id);
+        if (bookOptional.isPresent()) {
+            BookDto bookDto = BookDto.toDto(bookOptional.get());
+            model.addAttribute("bookDto", bookDto);
+        }
+        return "addbook";
     }
 }
