+++ dz14/src/test/java/ru/otus/dz14/controller/BookControllerTest.java
@@ -17,10 +17,12 @@ import ru.otus.dz14.service.LibraryService;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.Optional;
 
 import static org.hamcrest.Matchers.containsString;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 @RunWith(SpringRunner.class)
@@ -55,6 +57,7 @@ public class BookControllerTest {
         genre = new Genre("роман-эпопея");
         genre.setId(1);
         book = new Book("Война и мир", author, genre);
+        book.setId(1);
         comment = new Comment("Эпично, но слишком затянуто.", book);
         books = Arrays.asList(book);
         bookDto = new BookDto(1, "Мертвые души", "Николай", "Гоголь", "поэма");
@@ -87,5 +90,20 @@ public class BookControllerTest {
         mvc.perform(post("/books/add/1")
                 .flashAttr("bookDto", bookDto))
                 .andExpect(redirectedUrl("/books"));
+    }
+
+    @Test
+    public void addBookPage() throws Exception {
+        mvc.perform(get("/addbook"))
+                .andExpect(status().isOk());
+    }
+
+    @Test
+    public void editBookPage() throws Exception {
+        Mockito.when(libraryService.findBookById(1)).thenReturn(Optional.of(book));
+        mvc.perform(get("/addbook/edit?id=" + book.getId()))
+                .andExpect(status().isOk())
+                .andDo(print())
+                .andExpect(content().string(containsString(book.getTitle())));
     }
 }
